Copy oxygen and delay copy to second stage.
By delaying the copying of all files towards the second phase of the
container construction, the possibility to reuse the complete cache for
the first cache become higher. Thus, only the first stage is recomputed
when something changed in the specific requires and otherwise the cached
version is picked (avoiding preprocessing).