Add `clot-present` and `healthy-scenario` flags.
The first replaces the `clotactive` hardcocded value, such that we can
signal if the clot is present when invoking the script from the API.

The second indicates if the script should only run the initial
simulation, i.e. the healthy part, or also consider the two-way
coupling, which is only required when the clot is actually present.